Conflicts:
	Procfile
	Rakefile
	TODO.mkd
	app.py
	apps/accounts/handlers.py
	apps/accounts/models.py
	apps/accounts/social.py
	requirements.txt
	settings.py
	static/js/custom.js
	templates/accounts/register.html
	templates/accounts/user_page.html
	templates/base.html
	templates/main.html

+++ apps/utils/rest.py
@@ -1,11 +1,13 @@
 # coding: utf-8
-from apps.api import ApiHandler
+from .api import ApiHandler
 from apps.accounts.models import User # FIXME need refactoring
 
 
 class MongoEngineDataManager(object):
-    def __init__(self, model):
+    def __init__(self, model, user=None, ip=None):
         self.model = model
+        self.user = user
+        self.ip = ip
 
     def read_list(self, initial=0, amount=50):
         return self.model.objects.all()[initial:initial+amount]
@@ -42,10 +44,6 @@ class MongoEngineDataManager(object):
 
 
 class MongoEngineDataManagerPerUser(MongoEngineDataManager):
-    def __init__(self, model, user):
-        super(MongoEngineDataManagerPerUser, self).__init__(model)
-        self.user = user
-
     def read_list(self, initial=0, amount=50):
         objs = super(MongoEngineDataManagerPerUser, self).read_list(initial=initial, amount=amount)
         return objs(user=self.user)
@@ -59,8 +57,21 @@ class MongoEngineDataManagerPerUser(MongoEngineDataManager):
 
 
 class RestHandler(ApiHandler):
+    perm_public = 'CRLUD'
+    perm_user = 'CRLUD'
+    perm_admin = 'CRLUD'
+
+    def check_permission(self, action):
+        user = self.get_current_user()
+        admin = self.is_admin_user()
+        if action in self.perm_public or (user and action in self.perm_user) or (admin and action in self.perm_admin):
+            pass # ok
+        else:
+            self.raise403()
+
     # CREATE /objs
     def post(self):
+        self.check_permission('C')
         data = self.get_request_data()
         obj = self.data_manager.create(data)
         if obj:
@@ -72,6 +83,7 @@ class RestHandler(ApiHandler):
     # READ /objs/:id
     def get(self, identifier=None):
         if identifier:
+            self.check_permission('R')
             if identifier == 'count':
                 self.answer(self.data_manager.count())
             else:
@@ -81,6 +93,7 @@ class RestHandler(ApiHandler):
                 else:
                     self.raise404()
         else:
+            self.check_permission('L')
             # FIXME pagination
             initial = int(self.get_argument('initial', default=0))
             amount = int(self.get_argument('amount', default=50))
@@ -91,6 +104,7 @@ class RestHandler(ApiHandler):
 
     # UPDATE /objs/:id
     def put(self, identifier):
+        self.check_permission('U')
         data = self.get_request_data()
         obj = self.data_manager.update(identifier, data)
         if obj:
@@ -100,9 +114,9 @@ class RestHandler(ApiHandler):
 
     # DELETE /objs/:id
     def delete(self, identifier):
+        self.check_permission('D')
         obj = self.data_manager.delete(identifier)
         if obj:
             self.answer(obj)
         else:
             self.raise404()
-
